backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==1.4.49
pydantic==2.5.0
openai==1.3.7
python-dotenv==1.0.0
requests==2.31.0
jinja2==3.1.2
python-multipart==0.0.6
databases[sqlite]==0.8.0
aiosqlite==0.19.0 
httpx==0.24.1